robot-rl/pyproject.toml
[tool.poetry]
name = "robot-rl"
version = "0.1.0"
description = ""
authors = ["Ashis Ghosh <ahsisghosh@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
stable-baselines3 = "^2.3.2"
pybullet = "^3.2.6"
tensorboard = "^2.16.2"
gymnasium = "^0.29.1"
gymnasium-robotics = "^1.2.4"
wandb = "^0.17.0"
python-dotenv = "^1.0.1"
sb3-contrib = "^2.3.0"
moviepy = "^1.0.3"
hydra-core = "^1.3.2"
mani-skill = "^3.0.0b4"
tyro = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
